Confra/requirements.txt
streamlit
pandas
openpyxl
python-dotenv
gspread
oauth2client
supabase
sib-api-v3-sdk
scikit-learn
prophet
numpy
matplotlib
statsmodels
plotly
mercadopago